+++ c-ext/python-zstandard.h
@@ -15,6 +15,7 @@
 #include <zstd.h>
 #include <zdict.h>
 
+/* Remember to change the string in zstandard/__init__ as well */
 #define PYTHON_ZSTANDARD_VERSION "0.10.0.dev0"
 
 typedef enum {