+++ tests/test_module_attributes.py
@@ -14,6 +14,8 @@ class TestModuleAttributes(unittest.TestCase):
     def test_version(self):
         self.assertEqual(zstd.ZSTD_VERSION, (1, 3, 6))
 
+        self.assertEqual(zstd.__version__, '0.10.0.dev0')
+
     def test_constants(self):
         self.assertEqual(zstd.MAX_COMPRESSION_LEVEL, 22)
         self.assertEqual(zstd.FRAME_HEADER, b'\x28\xb5\x2f\xfd')