+++ zstandard/__init__.py
@@ -60,3 +60,5 @@ elif _module_policy == 'cffi':
 else:
     raise ImportError('unknown module import policy: %s; use default, cffi_fallback, '
                       'cext, or cffi' % _module_policy)
+
+__version__ = '0.10.0.dev0'