+++ NEWS.rst
@@ -88,6 +88,7 @@ New Features
 Changes
 -------
 
+* ``zstandard.__version__`` is now defined (#50).
 * Upgrade pip, setuptools, wheel, and cibuildwheel packages to latest versions.
 * Upgrade various packages used in CI to latest versions. Notably tox (in
   order to support Python 3.7).